+++ scripts/add_hmmsearch_to_gff.py
@@ -4,6 +4,45 @@ import sys
 import re
 import argparse
 import pandas
+import gffutils
+from signal import signal, SIGPIPE, SIG_DFL
+
+signal(SIGPIPE,SIG_DFL)
+
+def transcriptToGenome(db, txid, tx_start, tx_end, featuretype, parent_feature_type, child_feature_type):
+    features = list(db.children(txid, featuretype=featuretype, order_by='start'))
+    if len(features) == 0:
+        raise Exception(f'No feature of type "{featuretype}" found for ID "{txid}"')
+
+    tx = db[txid]
+    txToGenomeMap = {}
+    out = []
+    tx_pos = 1
+    for feature in features:
+        genome_pos = range(feature.start, feature.end + 1)
+        for i in genome_pos:
+            # We map tx coordinates to genome coordinates in a dictionary like:
+            # {1:11, 2:12, 3:100, 4:101, ...}
+            # That's horrible but easier later to query
+            txToGenomeMap[tx_pos] = i
+            tx_pos += 1
+        
+        if tx_start < tx_pos:
+            if txToGenomeMap[tx_start] < feature.start:
+                gstart = feature.start
+            else:
+                gstart = txToGenomeMap[tx_start] 
+            
+            if tx_end >= tx_pos:
+                gend = feature.end
+            else:
+                gend = txToGenomeMap[tx_end]
+            out.append(gffutils.Feature(tx.chrom, '.', child_feature_type, gstart, gend))
+            if tx_end < tx_pos:
+                break
+    parent = gffutils.Feature(tx.chrom, '.', parent_feature_type, out[0].start, out[-1].end)
+    out.insert(0, parent)
+    return(out)
 
 def read_tbldom(domfile):
     """Read domain file from hmmer (typically from `--domtblout`) and return a
@@ -13,26 +52,40 @@ def read_tbldom(domfile):
         fin = sys.stdin
     else:
         fin = open(domfile)
+    dom = fin.readlines()
+    if fin != '-':
+        fin.close()
+    
+    mode = [x for x in dom if x.startswith('# Pipeline mode:')][0]
+    if 'SEARCH' in mode:
+        mode = 'SEARCH'
+        colidx = [0, 3, 4]
+    elif 'SCAN' in mode:
+        mode = 'SCAN'
+        colidx = [3, 0, 1]
+    else:
+        raise Exception('Cannot determine type of hmmer output')
+    colidx.extend([12, 17, 18])
+
     tbldom = []
-    USECOLS = pandas.DataFrame({'colidx': [0, 3, 4, 12, 17, 18], 
+    USECOLS = pandas.DataFrame({'colidx': colidx, 
         'colname': ['protein_id', 'pfam_name', 'pfam_id', 'dom_evalue', 'aln_start', 'aln_end'], 
         'coltype': ['category', 'category', 'category', 'float64', 'int64', 'int64']})
-    for line in fin:
+    for line in dom:
         if line.startswith('#'):
             continue
         line = line.strip()
         line = re.sub(' +', '\t', line).split('\t')
-        if len(line) != 23:
+        if len(line) < 23:
             raise Exception('Error parsing line: "%s"' % ' '.join(line))
         line = [line[i] for i in USECOLS.colidx]
         tbldom.append(line)
-    if fin != '-':
-        fin.close()
     tbldom = pandas.DataFrame(tbldom, columns=USECOLS.colname)
     tbldom = tbldom.astype(dict(zip(USECOLS.colname, USECOLS.coltype)))
     tbldom.pfam_id = [x.split('.')[0] for x in tbldom.pfam_id]
     if len(tbldom) > 0:
         assert len(tbldom[tbldom.dom_evalue < 0]) == 0
+        assert all([x.startswith('PF') for x in tbldom.pfam_id])
     return tbldom
 
 def get_desc_from_hmm(hmmfile):
@@ -66,17 +119,9 @@ def get_desc_from_hmm(hmmfile):
     accdf.pfam_desc = list(acc2desc.values())
     return accdf
 
-def gff_encode(x):
-    """Replace reserved characters in `x` with URL encoding to make `x`
-    gff-compatible
-    """
-    x = re.sub(';', '%3B', x)
-    x = re.sub('=', '%3D', x)
-    x = re.sub(',', '%2C', x)
-    return x
-
-def read_pfam2go(pfam2go):
+def read_pfam2go(pfam2go, sep='|'):
     """Read pfam2go file (from sequenceontolgy.org) and return a DataFrame
+    sep: When a PFAM has multiple GOP terms, join terms using this separator
     """
     godf= pandas.DataFrame({'pfam_id': [], 'go_name': [], 'go_term': []}, dtype='category')
     if  pfam2go is None:
@@ -101,7 +146,6 @@ def read_pfam2go(pfam2go):
         go_name = line.split(' > ')[1].split(' ; ')[0].strip()
         assert go_name.startswith('GO:')
         go_name = re.sub('^GO:', '', go_name)
-        go_name = gff_encode(go_name)
         go_names.append(go_name)
 
         go_term = line.split(' ; ')[1]
@@ -112,27 +156,16 @@ def read_pfam2go(pfam2go):
     data = pandas.DataFrame({'pfam_id': pfam_ids, 'go_name': go_names, 'go_term': go_terms})
 
     godf = pandas.concat([godf, data])
-    godf = godf.groupby('pfam_id').agg({'go_term': lambda x: ','.join(x), 
-                                        'go_name': lambda x: ','.join(x)}).reset_index()
+    godf = godf.groupby('pfam_id').agg({'go_term': lambda x: sep.join(x), 
+                                        'go_name': lambda x: sep.join(x)}).reset_index()
     return godf
 
-def get_gff_key(gff_line, key):
-    """Extract attribute `key` from gff line
-    """
-    attr_str = gff_line.split('\t')[8]
-    attr = attr_str.rstrip(';').split(';')
-    for x in attr:
-        if x.startswith(key + '='):
-            x = x.replace(key + '=', '')
-            return x
-
-def dom2gff(tbldom, chrom, parent_start, parent_end, parent_strand, source='Pfam', feature='protein_match'):
+def dom2gff(tbldom, gff_db, source, parent_feature_type, child_feature_type):
     """Convert the rows in tbldom to GFF. tbldom is a DataFrame of hmmer hits
     for a single protein. Return these hits as a list of strings (lines) in GFF format. 
     
-    chrom, parent_start, parent_end, and parent_strand: Genomic coordinates of
-    the transcript producing the protein in question.
-    
+    gff_db: GFF database from gffutils.create_db()
+
     source and feature: Strings for columns 2 and 3 of output GFF. For
     "feature" use a valid sequence ontology term
     """
@@ -140,38 +173,55 @@ def dom2gff(tbldom, chrom, parent_start, parent_end, parent_strand, source='Pfam
     assert len(set(tbldom.protein_id)) in [0, 1]
     gff = []
     for idx, row in tbldom.iterrows():
-        if parent_strand == '+':
-            aln_start = parent_start + (row.aln_start - 1) * 3
-            aln_end = (parent_start + (row.aln_end - 1) * 3) - 1
-        elif parent_strand == '-':
-            aln_start = parent_end - (row.aln_end - 1) * 3 + 1
-            aln_end = parent_end - (row.aln_start - 1) * 3
-        assert aln_start < aln_end
-        assert aln_start >= parent_start
-        assert aln_end <= parent_end
         pfam_desc = row.pfam_desc
         if pandas.isna(pfam_desc):
             pfam_desc = 'NA'
-        pfam_desc = gff_encode(pfam_desc)
-        pfam_name = gff_encode(row.pfam_name)
+        pfam_desc = pfam_desc
+        pfam_name = row.pfam_name
         if pandas.isna(row.go_term):
-            go_ann = ''
+            ontology_name = 'NA'
+            Ontology_term = 'NA'
         else:
-            go_ann = f';ontology_name={row.go_name};Ontology_term={row.go_term}'
-        attr = f'ID={row.protein_id}.d{idx+1};Parent={row.protein_id};Name={pfam_name};pfam_id={row.pfam_id};dom_evalue={row.dom_evalue};signature_desc={pfam_desc}{go_ann}'
-        gff_row = [chrom, source, feature, aln_start, aln_end, row.dom_evalue, parent_strand, '.', attr]
-        gff_row = '\t'.join([str(x) for x in gff_row])
-        gff.append(gff_row)
+            ontology_name = row.go_name
+            Ontology_term = row.go_term
+        attr = [('ID', row.protein_id + '.d' + str(idx+1)),
+                ('Parent', row.protein_id),
+                ('Name', pfam_name),
+                ('pfam_id', row.pfam_id),
+                ('dom_evalue', row.dom_evalue),
+                ('signature_desc', pfam_desc),
+                ('ontology_name', ontology_name),
+                ('Ontology_term', Ontology_term)]
+        
+        hmm = transcriptToGenome(gff_db, row.protein_id, row.tx_start, row.tx_end, featuretype='CDS', parent_feature_type=parent_feature_type, child_feature_type=child_feature_type)
+
+        i = 1
+        pid = gff_db[row.protein_id]
+        for line in hmm:
+            line.source = source
+            line.strand = pid.strand
+            line.score = str(row.dom_evalue)
+            for k, v in attr:
+                line.attributes[k] = str(v)
+            if line.featuretype == parent_feature_type:
+                parent_id = line['ID']
+            if line.featuretype == child_feature_type:
+                line['ID'] = line['ID'][0] + '.' + str(i)
+                line['Parent'] = parent_id
+                i += 1
+            gff.append(line)
     return gff
 
 
-parser = argparse.ArgumentParser(description='Add the output of hmmsearch to gff. Typical usage: Get proteins from GFF, find protein domains (e.g. from Pfam) using hmmsearch, use this script to annotate the GFF with these domains')
+parser = argparse.ArgumentParser(description='Add the output of hmmsearch or hmmscan to gff. Typical usage: Get proteins from GFF, find protein domains (e.g. from Pfam) using hmmsearch, use this script to annotate the GFF with these domains')
 parser.add_argument('--gff', '-gff', help='GFF to annotate', required=True)
-parser.add_argument('--domtblout', '-d', help='Domain from `hmmsearch --domtblout <out>` [%(default)s]', required=True, default='-')
+parser.add_argument('--domtblout', '-d', help='Domain file from `hmmsearch or hmmscan --domtblout <out>` [%(default)s]', required=True, default='-')
 parser.add_argument('--hmm', '-H', help='Optional file of hmm profiles to extract domain descriptions, typically "Pfam-A.hmm". Use - to read from stdin [%(default)s]', default=None)
 parser.add_argument('--pfam2go', '-g', help='Optional file mapping Pfam domains to GO terms, typically "pfam2go" from http://www.sequenceontology.org/. Use - to read from stdin [%(default)s]', default=None)
 parser.add_argument('--evalue', '-e', help='Include domains with e-value below this cutoff [%(default)s]', default=0.01, type=float)
-parser.add_argument('--version', '-v', action='version', version='%(prog)s v0.2.0')
+parser.add_argument('--parent-feature-type', '-p', help='Set this feature type (column 3 in gff) for grouping spliced matches from the same hit [%(default)s]', default='protein_match')
+parser.add_argument('--child-feature-type', '-c', help='Set this feature type (column 3 in gff) for spliced matches [%(default)s]', default='protein_hmm_match')
+parser.add_argument('--version', '-v', action='version', version='%(prog)s v0.3.0')
 
 if __name__ == '__main__':
     args = parser.parse_args()
@@ -185,26 +235,24 @@ if __name__ == '__main__':
     tbldom = tbldom.merge(acc2desc, how='left', on='pfam_id')
     tbldom = tbldom.merge(pfam2go, how='left', on='pfam_id')
     assert n == len(tbldom)
-    
+
+    # Convert protein coordinates to transcript coordinates 
+    tbldom['tx_start'] = tbldom['aln_start'] * 3
+    tbldom['tx_end'] = tbldom['aln_end'] * 3
+
     if args.gff == '-':
         fin = sys.stdin
     else:
         fin = open(args.gff)
+    gff_db = gffutils.create_db(''.join(fin.readlines()), ":memory:", from_string=True)
 
     protein_ids = set(tbldom.protein_id)
-    first = True
-    for line in fin:
-        if first and not line.startswith('##gff-version'):
-            print('##gff-version 3')
-            first = False
-        if line.startswith('#'):
-            print(line)
-            continue
-        print(line.strip())
-        gff_id = get_gff_key(line, 'ID')
-        if gff_id in protein_ids:
-            hmm_hits = tbldom[tbldom.protein_id.isin([gff_id])]
-            gff_line = line.split('\t')
-            pfam_out = dom2gff(hmm_hits, chrom=gff_line[0], parent_start=int(gff_line[3]), parent_end=int(gff_line[4]), parent_strand=gff_line[6])
+    
+    print('##gff-version 3')
+    for feature in gff_db.all_features():
+        print(feature)
+        if feature.id in protein_ids:
+            hmm_hits = tbldom[tbldom.protein_id.isin([feature.id])]
+            pfam_out = dom2gff(hmm_hits, gff_db, source='Pfam', parent_feature_type=args.parent_feature_type, child_feature_type=args.child_feature_type)
             for out in pfam_out:
                 print(out)