+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 * Enable downloading protein database from URL
 
 * Remove `temp` tag from Pfam files
+
+* Fix mapping of Pfam domains spliced across exons
+